requirements.txt
appdirs==1.4.4
async-generator==1.10
attrs==21.2.0
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
colorama==0.4.4
cryptography==36.0.0
flake8==4.0.1
h11==0.12.0
idna==3.3
latest-user-agents==0.0.3
loguru==0.5.3
lxml==4.7.1
mccabe==0.6.1
numpy==1.21.5
outcome==1.1.0
pandas==1.3.5
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
pyOpenSSL==21.0.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
requests==2.26.0
schedule==1.1.0
selenium==4.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
trio==0.19.0
trio-websocket==0.9.2
urllib3==1.26.7
win32-setctime==1.0.4
wsproto==1.0.0
